websearch/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
python-dotenv
requests
numpy
fastmcp
pydantic
markitdown-mcp
